% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_uniprotIDs_to_MGIsymbols.r
\name{map_uniprotIDs_to_MGIsymbols}
\alias{map_uniprotIDs_to_MGIsymbols}
\title{map_uniprotIDs_to_MGIsymbols}
\usage{
map_uniprotIDs_to_MGIsymbols(uniprot_ids)
}
\arguments{
\item{uniprot_ids}{Array of uniprot IDs}
}
\value{
Array of gene symbols where each position equates to that in the inputs
}
\description{
\code{filter.genes.without.1to1.homolog} Takes an array of Uniprot IDs, returns the equivilent gene symbols
}
\examples{
map_uniprotIDs_to_MGIsymbols(uniprot_ids=c("Q05CX3","B1AR31","A1ILG8"))
}
