% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mgi_ids_2_syms}
\alias{mgi_ids_2_syms}
\title{MGI IDs to MGI symbols}
\format{An object of class \code{data.frame} with 49928 rows and 2 columns.}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
library("biomaRt")
mouse = useMart(host="www.ensembl.org", "ENSEMBL_MART_ENSEMBL", dataset = "mmusculus_gene_ensembl")
mgi_ids_2_syms = getBM(attributes=c("mgi_id","mgi_symbol"), mart=mouse)
devtools::use_data(mgi_ids_2_syms,overwrite=TRUE)
}
}
\usage{
mgi_ids_2_syms
}
\description{
A dataset obtained from Biomart mapping MGI IDs onto MGI symbols
}
\keyword{datasets}
