% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{uniprot_2_mgi}
\alias{uniprot_2_mgi}
\title{Uniprot 2 MGI}
\format{An object of class \code{data.frame} with 77102 rows and 9 columns.}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
library("biomaRt")
mouse = useMart(host="www.ensembl.org", "ENSEMBL_MART_ENSEMBL", dataset = "mmusculus_gene_ensembl")
attrib_mus = listAttributes(mouse)
mgi_symbols = getBM(attributes=c("mgi_id"), mart=mouse)
mgi_syms = getBM(attributes=c("mgi_id","mgi_symbol"), mart=mouse)
write.csv(mgi_syms,file="data/mgi_ids_to_syms.csv")
write.csv(mgi_symbols[,1],"mgi_ids.csv")
uniprot_2_mgi = fread("/Users/natske/Desktop/Mouse Synaptosome Project/data/uniprot-AllMGIids.tab",stringsAsFactors = FALSE)
uniprot_2_mgi = data.frame(uniprot_2_mgi)
colnames(uniprot_2_mgi)[1] = "MGIid"
devtools::use_data(uniprot_2_mgi,overwrite=TRUE)
}
}
\usage{
uniprot_2_mgi
}
\description{
A dataset obtained from Uniprot mapping each Swissprot/Trembl ID onto the equivilent MGI ID
}
\keyword{datasets}
